Completely move signal with start logic into its own package
